AEA-3830 Split validation code.
